test(telegram-api): share one harness across the nine callback suites
Each telegram-webhook-callbacks suite rebuilt the same ~100-line harness:
webhook env, D1 rows, update payloads and callback-query construction.
The clone detector measured 1,659 duplicated lines across the family,
the largest single clone family in the repository.

Extract it to telegram-webhook-callbacks.test-support.ts, beside the
existing telegram-webhook.test-support.ts, and migrate all nine suites.
Scaffolding only: every suite keeps its exact case count (69 cases total,
unchanged) and every assertion is untouched.

Net -584 lines.
