+++ worker/src/api/__tests__/telegram-webhook-callbacks-watchlist.test.ts
@@ -1,86 +1,16 @@
 import { describe, expect, it, vi, beforeEach } from "vitest";
-import { mockD1 as baseMockD1 } from "../../test-helpers/__shared/mock-d1";
 import {
-  createTelegramFetchSpy,
+  fetchSpy,
+  handleCallbackQuery,
+  mockD1,
+  resetCallbackTest,
+  sendAuditedTelegramReply,
+} from "./telegram-webhook-callbacks.test-support";
+import {
   telegramApiCallBody,
 } from "../../test-helpers/__shared/telegram";
 
-// Stub the insights module so /why callback tests do not need to drive the
-// full report-cards snapshot pipeline. Coverage callback uses buildCoverageMessage
-// which only reads StatusForCoin data, so it does not need stubbing.
-vi.mock("../telegram-webhook-insights", async (importOriginal) => {
-  const original = (await importOriginal()) as Record<string, unknown>;
-  return {
-    ...original,
-    buildWhyMessage: vi.fn(
-      async (_db: unknown, stablecoinId: string) => `<b>${stablecoinId} Safety Score</b>\nOverall: A`,
-    ),
-  };
-});
 
-// Indirect spy on sendAuditedTelegramReply so individual P1.16 tests can force
-// the reply to throw and assert the answer-callback still fires. The default
-// implementation defers to the real send path so all other tests keep working.
-vi.mock("../telegram-webhook-replies", async (importOriginal) => {
-  const original = (await importOriginal()) as Record<string, unknown>;
-  return {
-    ...original,
-    sendAuditedTelegramReply: vi.fn(original.sendAuditedTelegramReply as (...args: unknown[]) => Promise<void>),
-  };
-});
-
-const { handleCallbackQuery } = await import("../telegram-webhook-callbacks");
-const { sendAuditedTelegramReply } = await import("../telegram-webhook-replies");
-await import("../../lib/telegram-alerts");
-
-const { fetchSpy, reset: resetTelegramFetchSpy } = createTelegramFetchSpy();
-
-function mockD1(
-  tables: Parameters<typeof baseMockD1>[0] = [],
-  options: Parameters<typeof baseMockD1>[1] = {},
-) {
-  return baseMockD1([
-    ...tables,
-    { match: "FROM telegram_subscribers", rows: [], first: null },
-    { match: "FROM telegram_subscriptions", rows: [] },
-    { match: "FROM telegram_preset_subscriptions", rows: [] },
-    { match: "FROM telegram_pending_disambiguation", rows: [], first: null },
-    { match: "FROM telegram_pending_alerts", rows: [], first: null },
-    { match: "FROM telegram_recap_preferences", rows: [], first: null },
-    { match: "FROM cache", rows: [], first: null },
-    { match: "FROM price_cache", rows: [], first: null },
-    { match: "FROM dex_liquidity", rows: [], first: null },
-    { match: "FROM yield_data", rows: [], first: null },
-    { match: "FROM stress_signals", rows: [], first: null },
-    { match: "FROM stress_signal_publication_rows", rows: [], first: null },
-    { match: "FROM safety_grade_history", rows: [], first: null },
-    { match: "FROM depeg_events", rows: [], first: null },
-    { match: "INSERT INTO telegram_subscribers", rows: [] },
-    { match: "UPDATE telegram_subscribers", rows: [] },
-    { match: "INSERT INTO telegram_subscriptions", rows: [] },
-    { match: "UPDATE telegram_subscriptions", rows: [] },
-    { match: "DELETE FROM telegram_subscriptions", rows: [] },
-    { match: "INSERT INTO telegram_preset_subscriptions", rows: [] },
-    { match: "DELETE FROM telegram_preset_subscriptions", rows: [] },
-    { match: "INSERT INTO telegram_pending_disambiguation", rows: [] },
-    { match: "DELETE FROM telegram_pending_disambiguation", rows: [] },
-    { match: "DELETE FROM telegram_recap_targets", rows: [] },
-    { match: "UPDATE telegram_recap_preferences", rows: [] },
-    { match: "UPDATE telegram_recap_targets", rows: [] },
-    { match: "DELETE FROM telegram_pending_alerts", rows: [] },
-    { match: "DELETE FROM telegram_freeze_alert_targets", rows: [] },
-    { match: "DELETE FROM telegram_alert_source_resolution_targets", rows: [] },
-    { match: "DELETE FROM telegram_alert_target_plan_items", rows: [] },
-    { match: "DELETE FROM telegram_alert_job_targets", rows: [] },
-    { match: "DELETE FROM telegram_alert_job_target_items", rows: [] },
-    { match: "DELETE FROM telegram_alert_target_plans", rows: [] },
-    { match: "DELETE FROM telegram_alert_planning_subscribers", rows: [] },
-    { match: "DELETE FROM telegram_transport_failure_observations", rows: [] },
-    { match: "DELETE FROM telegram_alert_dead_letters", rows: [] },
-    { match: "DELETE FROM telegram_chat_delivery_diagnostics", rows: [] },
-    { match: "INSERT INTO telegram_usage_daily", rows: [] },
-  ], options);
-}
 
 
 
@@ -100,12 +30,7 @@ function firstAckBody(): { text?: string } {
 
 
 
-beforeEach(() => {
-  resetTelegramFetchSpy();
-  // mockRejectedValueOnce expires after a single call so subsequent tests reuse
-  // the original sendAuditedTelegramReply impl bound at vi.mock construction.
-  vi.mocked(sendAuditedTelegramReply).mockClear();
-});
+beforeEach(resetCallbackTest);
 
 describe("handleCallbackQuery", () => {
   describe("watchlist manage (P1-U8)", () => {
